+++ src/javascript/binary/websocket_pages/trade/beta/analysis.js
@@ -1,4 +1,5 @@
 var AssetIndexUI = require('../../resources/asset_index/asset_indexws.ui').AssetIndexUI;
+var MarketTimesUI = require('../../resources/market_times/market_timesws.ui').MarketTimesUI;
 
 /*
  * This file contains the code related to loading of trading page bottom analysis
