+++ src/javascript/binary/websocket_pages/trade/beta/message.js
@@ -1,4 +1,5 @@
 var AssetIndexUI = require('../../resources/asset_index/asset_indexws.ui').AssetIndexUI;
+var MarketTimesUI = require('../../resources/market_times/market_timesws.ui').MarketTimesUI;
 
 /*
  * This Message object process the response from server and fire
